moved state func to common
