+++ internal/endpoints/mobiledeviceconfigurationprofilesplist/mobiledeviceconfigurationprofilesplist_crud.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
-	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/state"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -72,7 +72,7 @@ func resourceJamfProMobileDeviceConfigurationProfilePlistRead(ctx context.Contex
 	})
 
 	if err != nil {
-		return append(diags, state.HandleResourceNotFoundError(err, d, cleanup)...)
+		return append(diags, common.HandleResourceNotFoundError(err, d, cleanup)...)
 	}
 
 	return append(diags, updateTerraformState(d, response)...)
